Update for utils refactor
