+++ covid_model_parametrization/vulnerability.py
@@ -2,7 +2,6 @@
 
 import os
 import glob
-import argparse
 import logging
 from pathlib import Path
 import ast
@@ -18,8 +17,7 @@ from rasterio.mask import mask
 import fiona
 
 from covid_model_parametrization.exposure import get_output_filename
-
-from covid_model_parametrization import utils
+from covid_model_parametrization.utils import utils
 from covid_model_parametrization.config import Config
